--- a/ecc/bls12-377/fp/element_ops_amd64.s
+++ b/ecc/bls12-377/fp/element_ops_amd64.s
@@ -1,3 +1,5 @@
+// +build !purego
+
 // Copyright 2020 ConsenSys Software Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");