--- a/ecc/bls12-377/fp/element_ops_purego.go
+++ b/ecc/bls12-377/fp/element_ops_purego.go
@@ -1,5 +1,5 @@
-//go:build !amd64
-// +build !amd64
+//go:build !amd64 || purego
+// +build !amd64 purego
 
 // Copyright 2020 ConsenSys Software Inc.
 //