--- a/ecc/bls12-377/fp/element_ops_purego.go
+++ b/ecc/bls12-377/fp/element_ops_purego.go
@@ -46,14 +46,6 @@ func MulBy13(x *Element) {
 	x.Mul(x, &y)
 }
 
-// Butterfly sets
-//
-//	a = a + b (mod q)
-//	b = a - b (mod q)
-func Butterfly(a, b *Element) {
-	_butterflyGeneric(a, b)
-}
-
 func fromMont(z *Element) {
 	_fromMontGeneric(z)
 }