```
// Butterfly sets
//
//	a = a + b (mod q)
//	b = a - b (mod q)
//
func Butterfly(a, b *Element)
```

Is important for performance of the FFT and, more recently, for the MSM in Twisted edwards coordinates (see https://eprint.iacr.org/2022/1400.pdf ) . 

`gnark-crypto` has a `x86` impl in assembly, but not in `arm64`. This PR fixes that, BUT, not sure to merge:

1. This would be the only piece of `arm64` assembly, not sure it's needed for now.
2. The performance impact (wrt. compiler inlining) is not clear; for example, bls12-381/fr/fft impact: 

```
name                     old time/op  new time/op  delta
FFTDITCosetReference-10  51.0ms ± 2%  34.1ms ± 4%  -33.07%  (p=0.008 n=5+5)
FFTDIFReference-10       29.0ms ± 2%  36.3ms ± 2%  +25.36%  (p=0.008 n=5+5)
```

(on M1, need to bench on other more reliable ARM targets).

Keeping this as DRAFT for now in case in come handy in the future -- .